planning: move trajectory1d to common; deprecate unused classes
