+++ modules/planning/common/trajectory1d/piecewise_acceleration_trajectory1d.h
