+++ modules/planning/common/trajectory1d/piecewise_jerk_trajectory1d.cc
@@ -20,7 +20,7 @@
 
 #include <algorithm>
 
-#include "modules/planning/lattice/trajectory1d/piecewise_jerk_trajectory1d.h"
+#include "modules/planning/common/trajectory1d/piecewise_jerk_trajectory1d.h"
 #include "cyber/common/log.h"
 #include "modules/planning/common/planning_gflags.h"
