+++ modules/planning/common/trajectory1d/piecewise_jerk_trajectory1d.h
@@ -23,7 +23,7 @@
 #include <string>
 #include <vector>
 
-#include "modules/planning/lattice/trajectory1d/constant_jerk_trajectory1d.h"
+#include "modules/planning/common/trajectory1d/constant_jerk_trajectory1d.h"
 #include "modules/planning/math/curve1d/curve1d.h"
 
 namespace apollo {
