+++ modules/planning/common/trajectory1d/piecewise_trajectory1d.h
