+++ modules/planning/lattice/trajectory_generation/lateral_osqp_optimizer.h
@@ -25,7 +25,7 @@
 
 #include "osqp/include/osqp.h"
 
-#include "modules/planning/lattice/trajectory1d/piecewise_jerk_trajectory1d.h"
+#include "modules/planning/common/trajectory1d/piecewise_jerk_trajectory1d.h"
 #include "modules/planning/lattice/trajectory_generation/lateral_qp_optimizer.h"
 
 namespace apollo {
