+++ modules/planning/lattice/trajectory_generation/lattice_trajectory1d.h
