+++ modules/planning/lattice/trajectory_generation/piecewise_braking_trajectory_generator.h
@@ -22,7 +22,7 @@
 
 #include <memory>
 
-#include "modules/planning/lattice/trajectory1d/piecewise_acceleration_trajectory1d.h"
+#include "modules/planning/common/trajectory1d/piecewise_acceleration_trajectory1d.h"
 
 namespace apollo {
 namespace planning {
