+++ modules/planning/lattice/trajectory_generation/trajectory1d_generator.h
@@ -28,7 +28,7 @@
 
 #include "modules/planning/lattice/behavior/path_time_graph.h"
 #include "modules/planning/lattice/behavior/prediction_querier.h"
-#include "modules/planning/lattice/trajectory1d/lattice_trajectory1d.h"
+#include "modules/planning/lattice/trajectory_generation/lattice_trajectory1d.h"
 #include "modules/planning/lattice/trajectory_generation/end_condition_sampler.h"
 #include "modules/planning/math/curve1d/curve1d.h"
 #include "modules/planning/math/curve1d/quartic_polynomial_curve1d.h"
