+++ modules/planning/lattice/trajectory_generation/trajectory_evaluator.cc
@@ -27,7 +27,7 @@
 #include "modules/common/math/path_matcher.h"
 #include "modules/planning/common/planning_gflags.h"
 #include "modules/planning/constraint_checker/constraint_checker1d.h"
-#include "modules/planning/lattice/trajectory1d/piecewise_acceleration_trajectory1d.h"
+#include "modules/planning/common/trajectory1d/piecewise_acceleration_trajectory1d.h"
 #include "modules/planning/lattice/trajectory_generation/piecewise_braking_trajectory_generator.h"
 
 namespace apollo {
