+++ modules/planning/planner/lattice/BUILD
@@ -19,7 +19,7 @@ cc_library(
         "//modules/planning/constraint_checker",
         "//modules/planning/constraint_checker:collision_checker",
         "//modules/planning/lattice/behavior:path_time_graph",
-        "//modules/planning/lattice/trajectory1d:lattice_trajectory1d",
+        "//modules/planning/lattice/trajectory_generation:lattice_trajectory1d",
         "//modules/planning/lattice/trajectory_generation:backup_trajectory_generator",
         "//modules/planning/lattice/trajectory_generation:trajectory1d_generator",
         "//modules/planning/lattice/trajectory_generation:trajectory_combiner",
