+++ modules/planning/planner/lattice/lattice_planner.cc
@@ -35,7 +35,7 @@
 #include "modules/planning/constraint_checker/constraint_checker.h"
 #include "modules/planning/lattice/behavior/path_time_graph.h"
 #include "modules/planning/lattice/behavior/prediction_querier.h"
-#include "modules/planning/lattice/trajectory1d/lattice_trajectory1d.h"
+#include "modules/planning/lattice/trajectory_generation/lattice_trajectory1d.h"
 #include "modules/planning/lattice/trajectory_generation/backup_trajectory_generator.h"
 #include "modules/planning/lattice/trajectory_generation/trajectory1d_generator.h"
 #include "modules/planning/lattice/trajectory_generation/trajectory_combiner.h"
